==> projects/kdd/1_selection.ipynb <==
# 1. Selecão

from ucimlrepo import fetch_ucirepo

adult = fetch_ucirepo(id=2)

df = adult.data.features
df.head()

target = adult.data.targets
target.head()

df["income"] = target

df.to_csv("./data/raw/adult_raw.csv", index=False)
